아이템과 마을사람의 y값이 겹치지 않도록 수정

+++ Assets/Scripts/ItemBase.cs
@@ -19,7 +19,7 @@ public class ItemBase : MonoBehaviour
             gameObject.SetActive(true);
         }
         #region position X
-        LineNum = Random.Range(0, 3);
+        LineNum = (int)Random.Range(0, 3);
         if(LineNum == 0)
         {
             posX = -1.4f;
